RFCT(Mark final sets of tests as requiring windows)
