+++ src/test/java/mmarquee/automation/controls/ribbon/AutomationRibbonCommandBarTest.java
@@ -74,61 +74,4 @@ public class AutomationRibbonCommandBarTest {
 
         assertTrue(name.equals("NAME"));
     }
-
-    @Test (expected = ElementNotFoundException.class)
-    public void testGetRibbonCommandBar_Throws_Exception_When_No_WorkPane_Found() throws Exception {
-        AutomationElement element = Mockito.mock(AutomationElement.class);
-
-        List<AutomationElement> collection = new ArrayList<>();
-
-        when(element.getClassName()).thenReturn(AutomationRibbonCommandBar.CLASS_NAME);
-        when(element.findAll(any(), any())).thenReturn(collection);
-
-        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
-        UIAutomation instance = new UIAutomation(mocked_automation);
-
-        ItemContainer container = Mockito.mock(ItemContainer.class);
-
-        AutomationRibbonCommandBar commandBar = new AutomationRibbonCommandBar(element, container, instance);
-
-        AutomationRibbonWorkPane workPane = commandBar.getRibbonWorkPane();
-
-        Mockito.verify(element, atLeastOnce()).findAll(any(), any());
-    }
-
-    @Test
-    @Ignore("Requires Windows")
-    public void testGetRibbonCommandBar_When_WorkPane_Found() throws Exception {
-        AutomationElement element = Mockito.mock(AutomationElement.class);
-
-        List<AutomationElement> collection = new ArrayList<>();
-
-        IUIAutomationElement3 elem = Mockito.mock(IUIAutomationElement3.class);
-
-        Mockito.when(elem.getCurrentClassName(any())).thenAnswer(
-                invocation -> {
-                    Object[] args = invocation.getArguments();
-                    PointerByReference reference = (PointerByReference) args[0];
-
-                    String value = "UIRibbonWorkPane";
-                    Pointer pointer = new Memory(Native.WCHAR_SIZE * (value.length() + 1));
-                    pointer.setWideString(0, value);
-
-                    reference.setValue(pointer);
-
-                    return 0;
-                }
-        );
-
-        collection.add(new AutomationElement(elem));
-
-        when(element.getClassName()).thenReturn(AutomationRibbonCommandBar.CLASS_NAME);
-        when(element.findAll(any(), any())).thenReturn(collection);
-
-        AutomationRibbonCommandBar commandBar = new AutomationRibbonCommandBar(element);
-
-        AutomationRibbonWorkPane workPane = commandBar.getRibbonWorkPane();
-
-        Mockito.verify(element, atLeastOnce()).findAll(any(), any());
-    }
 }
